requirements.txt
numpy
matplotlib
torchvision
pickle
gensim
pretrainedmodels